# bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "DealerPro",
    "version": "v1"
  },
  "paths": {
...
    "/api/DashBoardData/userid": {
      "post": {
        "tags": [
          "DashBoardData"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TokenData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/DashBoardData/create": {
      "post": {
        "tags": [
          "DashBoardData"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/storeData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/DashBoardData/changeName": {
      "put": {
        "tags": [
          "DashBoardData"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UID",
            "type": "string"
          },
          {
            "in": "query",
            "name": "existStoreName",
            "type": "string"
          },
          {
            "in": "query",
            "name": "newStoreName",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/DashBoardData/deleteStore": {
      "delete": {
        "tags": [
          "DashBoardData"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UID",
            "type": "string"
          },
          {
            "in": "query",
            "name": "storeName",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/DashBoardData/getStoreName": {
      "get": {
        "tags": [
          "DashBoardData"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "Token",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/DashBoardData/CheckForStoreDuplicate": {
      "post": {
        "tags": [
          "DashBoardData"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "TokenData": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        }
      }
    },
...
    "storeData": {
      "type": "object",
      "properties": {
        "uid": {
          "type": "string"
        },
        "storeName": {
          "type": "string"
        }
      }
    }
  }
}